move injectable functions to module
scripts would import from here in mv2
